Finalizar a parte de exclusão e reenvio no painel de reprovadas, edição já está completa

+++ routes/blueprint_painel_solicitacoes.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, render_template, request, redirect, url_for, flash
 from flask_login import login_required, current_user
-from models.database import Solicitacoes
+from models.database import Solicitacoes, db
 
 blueprint_painel_solicitacoes = Blueprint('blueprint_painel_solicitacoes', __name__)
 
@@ -21,20 +21,41 @@ def painel_solicitacoes():
                 query = query.order_by(Solicitacoes.STATUS)
         return render_template('painel_solicitacoes.html', solicitacoes=query, filtro=filtro, usuario_logado=current_user)
 
-@blueprint_painel_solicitacoes.route('/excluir-solicitacao/<int:id>', methods=['POST', 'GET'])
-@login_required
-def excluir_solicitacao(id):
-        solicitacao = Solicitacoes.get_or_none(Solicitacoes.id == id)
-
-        if request.method == 'POST':
-                solicitacao_deletada = Solicitacoes.get(Solicitacoes.id == id)
-                solicitacao_deletada.delete_instance()
-                flash('Solicitação deletada com sucesso!')
-                return redirect(url_for('blueprint_painel_solicitacoes.painel_solicitacoes', usuario_logado=current_user.USUARIO))
-
-
 @blueprint_painel_solicitacoes.route('/mais_info_sol/<int:id>')
 @login_required
 def mais_info_sol(id):
         solicitacao = Solicitacoes.get_or_none(Solicitacoes.id == id)
-        return render_template('mais_info_sol.html', usuario_logado=current_user.USUARIO, solicitacao=solicitacao)
+        return render_template('mais_info_sol.html', usuario_logado=current_user, solicitacao=solicitacao)
+
+
+@blueprint_painel_solicitacoes.route('/mais_info_sol/<int:id>/salvar', methods=['POST', 'GET'])
+@login_required
+def salvar_edicao(id):
+        solicitacao = Solicitacoes.get_or_none(Solicitacoes.id == id)
+        if not solicitacao:
+                flash('Solicitação não foi encontrada!', 'error')
+                return redirect(url_for('blueprint_painel_solicitacoes.painel_solicitacoes'))
+                
+        solicitacao.DEPARTAMENTO = request.form.get('departamento')
+        solicitacao.TIPO_DESPESA = request.form.get('tipo_despesa')
+        solicitacao.DESCRICAO = request.form.get('descricao')
+        solicitacao.VALOR = float(request.form.get('valor').replace('R$', '').replace('.', '').replace(',', '.').strip())
+
+        solicitacao.save() 
+        db.commit()
+        flash('Alterações realizadas com sucesso!', 'success')
+        return redirect(url_for('blueprint_painel_solicitacoes.mais_info_sol', id=id))
+
+@blueprint_painel_solicitacoes.route('/mais_info_sol/<int:id>/excluir', methods=['POST', 'GET'])
+@login_required
+def excluir_solicitacao(id):
+        if request.method == 'POST':
+                solicitacao = Solicitacoes.get_or_none(Solicitacoes.id == id)
+                if not solicitacao:
+                        flash('Nenhuma Solicitação foi encontrada!', 'error')
+                        return redirect(url_for('blueprint_painel_solicitacoes.mais_info_sol', usuario_logado=current_user))
+                solicitacao_deletada = Solicitacoes.get(Solicitacoes.id == id)
+                solicitacao_deletada.delete_instance()
+                flash('Solicitação Excluida com sucesso!', 'success')
+                return redirect(url_for('blueprint_painel_solicitacoes.mais_info_sol', usuario_logado=current_user, id=id))
+        return render_template('mais_info_sol.html', usuario_logado=current_user)
